+++ tests/stream/SeekTest.php
@@ -8,17 +8,25 @@ class SeekTest extends \voilab\csv\test\Seek
     protected function setUp() : void
     {
         parent::setUp();
-        $this->resource = $this->getResource('csv-seek.csv', true);
+        $this->resource = $this->getResource('csv-seek.csv');
     }
 
     public function testSeekWithStartUnsync() : void
     {
-        $this->markTestIncomplete('Known issue. Stream and seek does not work together.');
-    }
-
-    public function testSeek() : void
-    {
-        $this->markTestIncomplete('Known issue. Stream and seek does not work together.');
+        $result = $this->parser->parse($this->resource, [
+            'start' => 0,
+            'size' => 3
+        ]);
+        $result = $this->parser->parse($this->resource, [
+            'size' => 3,
+            'seek' => $this->resource->tell()
+        ]);
+        $expect = [
+            [ 'A' => 4, 'B' => 'test 4' ],
+            [ 'A' => 5, 'B' => 'test 5' ],
+            [ 'A' => 6, 'B' => 'test 6' ]
+        ];
+        $this->assertEquals($result, $expect);
     }
 
     public function testSeekPointerOverflow() : void